Update to use newer Vundle. Bit of cleanup as well.
Something strange keeps happening with my vim -in-terminal (as opposed
to something like gvim or MacVim), where the first input is taking as a
delete instruction, as if 'd' was pressed when the file was opened.
Sometimes I also find it's entered replace mode. Would like to figure it
out. Weird.